+++ core-rust/mesh-api-server/src/mesh_api/conversions/block.rs
@@ -1,0 +1,20 @@
+use crate::prelude::*;
+
+/// We assume that Block is a single transaction.
+/// Block index => State version
+/// Block hash  => State version printed to string and prefixed with zeros
+pub(crate) fn to_mesh_api_block_identifier_from_state_version(
+    state_version: StateVersion,
+) -> Result<models::BlockIdentifier, MappingError> {
+    let index = to_mesh_api_block_index_from_state_version(state_version)?;
+    Ok(models::BlockIdentifier {
+        index,
+        hash: format!("{:0>32}", index),
+    })
+}
+
+pub(crate) fn to_mesh_api_block_identifier_from_ledger_header(
+    ledger_header: &LedgerStateSummary,
+) -> Result<models::BlockIdentifier, MappingError> {
+    to_mesh_api_block_identifier_from_state_version(ledger_header.state_version)
+}
