+++ core-rust/mesh-api-server/src/mesh_api/conversions/mod.rs
@@ -1,4 +1,5 @@
 mod addressing;
+mod block;
 mod common;
 mod context;
 mod errors;
@@ -6,6 +7,7 @@ mod hashes;
 mod numerics;
 
 pub use addressing::*;
+pub use block::*;
 pub use common::*;
 pub use context::*;
 pub use errors::*;
