+++ core-rust/mesh-api-server/src/mesh_api/helpers.rs
@@ -66,42 +66,3 @@ pub(crate) fn read_optional_substate<D: ScryptoDecode>(
 ) -> Option<D> {
     database.get_substate::<D>(node_id, partition_number, substate_key)
 }
-
-/// We assume that Block is a single transaction.
-/// Block index => State version
-/// Block hash  => State version printed to string and prefixed with zeros
-pub(crate) fn to_mesh_api_block_identifier_from_state_version(
-    state_version: StateVersion,
-) -> Result<models::BlockIdentifier, MappingError> {
-    let index = to_mesh_api_block_index_from_state_version(state_version)?;
-    Ok(models::BlockIdentifier {
-        index,
-        hash: format!("{:0>32}", index),
-    })
-}
-
-pub(crate) fn to_mesh_api_block_identifier_from_ledger_header(
-    ledger_header: &LedgerStateSummary,
-) -> Result<models::BlockIdentifier, MappingError> {
-    to_mesh_api_block_identifier_from_state_version(ledger_header.state_version)
-}
-
-pub(crate) fn extract_state_version_from_mesh_api_partial_block_identifier(
-    block_identifier: &models::PartialBlockIdentifier,
-) -> Result<Option<StateVersion>, ExtractionError> {
-    let state_version = if let Some(index) = block_identifier.index {
-        Some(index)
-    } else if let Some(hash) = &block_identifier.hash {
-        Some(
-            hash.parse::<i64>()
-                .map_err(|_| ExtractionError::InvalidBlockIdentifier {
-                    message: "Hash parsing error".to_string(),
-                })?,
-        )
-    } else {
-        None
-    }
-    .map(|index| StateVersion::of(index as u64));
-
-    Ok(state_version)
-}
